requirements.txt
numpy
pandas
tensorflow
scikit-learn
matplotlib
seaborn
pyyaml
instaloader